Modifies prefix and suffix diffs to not split UTF-8
A diff splitting a UTF-8 codepoint is possible: Ω and ϩ, for example,
share a common suffix.  This split will make the diff invalid UTF-8,
even if both texts are valid.  A greater concern is common prefixes:
every Greek letter between Α and ο shares the prefix byte 0xce, meaning
that splits on a prefix would be the rule, not the exception, in normal
multibyte texts.

This patch prevents the prefix and suffix diffing from ever splitting
on such common prefixes and suffixes, by backing out when a diff lands
on a follow byte.  It remains to ensure that other splits are never in
the midst of a multibyte codepoint.

Signed-off-by: Sam Atman <atmanistan@gmail.com>