Add equalForwards and equalBackwards for Myers 1986
The only thing which matters here is not splitting codepoints down the
middle.  Since prefix and suffix matching will also refuse to split,
later repair stages won't re-split on us.  As the comments note, this
doesn't validate UTF-8, and might create less-than-perfect diffs if the
texts contain bad Unicode, a situation and outcome I do not care about
even slightly.

In addition to fixing the problem where diffs would be invalid utf-8 in
many ordinary cases, what this gets us is the ability to use codepoints
as indices for the diffLine function.  Since diffing won't damage utf-8
sequences, we can diff in linemode and decode each sequence back to the
codepoint, which will serve as an offset into the line_array.

Signed-off-by: Sam Atman <atmanistan@gmail.com>