Add match section
It's number-crunchy code, but it doesn't do a lot of allocation.  There
may be typos, it would surprise me if there are none in fact, but other
than that, it should work fine.

That leaves adding patchApply, translating enough unit tests to achieve
full coverage, and then refactoring the surface area to be a more ergo-
nomic.

Signed-off-by: Sam Atman <atmanistan@gmail.com>